requirements.txt
# Core dependencies for AI Tutor application

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Templating
jinja2==3.1.2

# LLM Providers
# OpenAI API (optional - only needed if using OpenAI)
# Version 2.15.0+ required for Responses API support
openai>=2.15.0

# HTTP requests for Ollama API
requests==2.31.0

# Scientific computing
numpy==1.26.2

# Knowledge Base Building (Hugging Face embeddings)
sentence-transformers==2.2.2

# Environment management
python-dotenv==1.0.0

# Frontend enhancements (loaded via CDN, but listed for reference)
# marked.js - Markdown rendering (CDN)
# highlight.js - Syntax highlighting (CDN)
# dompurify - XSS protection (CDN)

# Optional: Development dependencies
# Uncomment if needed for development

# Code formatting
# black==23.11.0
# isort==5.12.0

# Testing
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2
